+++ CHANGELOG.md
@@ -24,3 +24,4 @@ This change log follows the conventions of [keepachangelog.com](http://keepachan
 - `build` is now on Github.
 - `puml` files were missing plantuml directives
 - process complete directory; Both SVG and PUML output are optional
+- do not generate unchanged file